Apply rustfmt + Clippy fixes across all crates for consistent style
Run rustfmt across the entire workspace to normalize formatting that
had drifted as new code was added. This makes future diffs cleaner and
enforces a single canonical style.

Key formatting changes:
- Reorganize import blocks in ft-dispatch to one-per-line alphabetical
  order (previously packed into dense multi-import lines that exceeded
  the line width limit)
- Unwrap short function signatures onto single lines where they fit
  (to_f32, to_f64, unary_contiguous_f32, named_modules, etc.)
- Reflow long function call arguments to multi-line when they exceed
  the column limit (argmax_dim, argmin_dim, max_dim, min_dim,
  masked_fill, reduce_sum_for_broadcast, GroupNorm eps, Conv2d padding)
- Collapse short assert! and assert_eq! bodies onto single lines, and
  expand long assertion messages to multi-line format
- Normalize tuple return formatting in autograd tensor operations
- Reformat erf_value_f32 float literals with underscore separators per
  Rust style conventions (0.254829592f32 -> 0.254_829_6_f32)

Semantic fixes caught by Clippy:
- ft-nn: replace manual range check `v >= -1.0 && v <= 1.0` with
  idiomatic `(-1.0..=1.0).contains(&v)` (clippy::manual_range_contains)
- ft-optim: remove redundant `as i32` casts in StepLR/ExponentialLR
  powi() calls where the epoch variable already has the correct type

No behavioral changes beyond the two Clippy fixes above.

+++ crates/ft-optim/src/lib.rs
@@ -727,8 +727,7 @@ impl Optimizer for RMSprop {
             let avg: Vec<f64> = if self.centered {
                 // Update running average of gradients:
                 // grad_avg = alpha * grad_avg + (1 - alpha) * grad
-                let ga =
-                    self.grad_avg[i].get_or_insert_with(|| vec![0.0; effective_grad.len()]);
+                let ga = self.grad_avg[i].get_or_insert_with(|| vec![0.0; effective_grad.len()]);
                 ensure_state_len(
                     effective_grad.len(),
                     ga.len(),
@@ -748,8 +747,8 @@ impl Optimizer for RMSprop {
 
             if self.momentum > 0.0 {
                 // Update momentum buffer: buf = momentum * buf + grad / avg
-                let buf = self.momentum_buffer[i]
-                    .get_or_insert_with(|| vec![0.0; effective_grad.len()]);
+                let buf =
+                    self.momentum_buffer[i].get_or_insert_with(|| vec![0.0; effective_grad.len()]);
                 ensure_state_len(
                     effective_grad.len(),
                     buf.len(),
@@ -3239,7 +3238,7 @@ mod tests {
 
         for epoch in 0..30 {
             scheduler.step(&mut opt, None);
-            let expected = 0.5_f64.powi((epoch as i32) / 10);
+            let expected = 0.5_f64.powi(epoch / 10);
             let actual = opt.get_lr();
             assert!(
                 (actual - expected).abs() < 1e-12,
@@ -3303,7 +3302,7 @@ mod tests {
 
         for epoch in 0..5 {
             scheduler.step(&mut opt, None);
-            let expected = 0.5_f64.powi(epoch as i32);
+            let expected = 0.5_f64.powi(epoch);
             let actual = opt.get_lr();
             assert!(
                 (actual - expected).abs() < 1e-12,
